refactor: scrap페이지 비즈니스 로직 분리
