+++ src/pages/scrap/[docId].tsx
@@ -2,57 +2,29 @@ import { MoreVert as MoreVertIcon } from '@mui/icons-material';
 import { PushPin as PushPinIcon } from '@mui/icons-material';
 import { Button, ListItem, ListItemText, Menu, MenuItem, Paper } from '@mui/material';
 import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
-import { deleteScrap, getScrapDetail } from '@/api/scrap';
 import { ScrapList } from '@/components/scrap';
+import useScrapDetailData from '@/hooks/scrap/useScrapDetailData';
 import { scrapListTextStyle, scrapPaperStyle } from '@/styles/commonStyle';
-import { ScrapDetailType } from '@/types/scrap';
 
 const ScrapDetailPage = () => {
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const [scrapId, setScrapId] = useState<string>('');
-  const [scrapTitle, setScrapTitle] = useState();
-  const [scrapContents, setScrapContents] = useState<ScrapDetailType[]>();
-  const [postId, setPostId] = useState<string>('');
   const router = useRouter();
-  const open = Boolean(anchorEl);
+  const {
+    anchorEl,
+    scrapTitle,
+    scrapContents,
+    postId,
+    open,
+    fetchScrap,
+    handleClick,
+    handleClose,
+    deleteScrapItem,
+  } = useScrapDetailData();
 
   useEffect(() => {
-    const fetchScrapDoc = async () => {
-      if (router.isReady && typeof router.query.docId === 'string') {
-        const { data } = await getScrapDetail(router.query.docId);
-        setScrapTitle(data.title);
-        setScrapContents(data.contents);
-      }
-    };
-    fetchScrapDoc();
-  }, [router]);
-
-  const handleClick = (
-    event: React.MouseEvent<HTMLButtonElement>,
-    scrapId: string,
-    postId: string,
-  ) => {
-    setScrapId(scrapId);
-    setPostId(postId);
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const deleteScrapItem = async () => {
-    handleClose();
-    if (router.isReady && typeof router.query.docId === 'string') {
-      deleteScrap(router.query.docId, scrapId);
-      setScrapContents(
-        (prevContents) =>
-          prevContents &&
-          prevContents.filter((content) => content.scrapObjectId !== scrapId),
-      );
-    }
-  };
+    fetchScrap();
+  }, [fetchScrap]);
 
   const handleRouter = () => {
     router.push({
